refactor: squash migrations 003-009 into one
None of the seven incremental migrations built up during this PR's
development were ever applied to a real database, so the intermediate
states they captured (adding events.allow_preorders in one migration
and dropping it again one migration later; three separate steps for
amount_due, the external-partner-to-co-organiser swap, and products;
two more for required/bundling) had no reason to survive as separate
migration files once the PR is squash-merged.

Consolidated into a single 003, verified to apply and fully downgrade
against a real Postgres instance. allow_preorders never appears at
all, since it was added and dropped within the same unreleased span.
Commit history (and this PR's description) still narrate the design
evolution; the migration chain now reflects only the end state.
